requirements.txt
Flask
Flask-JWT
Flask-SQLAlchemy
mysqlclient
flasgger
Flask-Cors
bcrypt
cryptography
pysqlite3